v0.12.6
 (done) таск 12. баг непоследовательности сообщений в сопряжении (добавляется не на последнее место в чате)
            (если удалять крайние сообщения - выглядит так, буд-то счётчик откатывается на пару сообщений назад
 таск 9. подружить синхронное удаление сообщений (кнопка "видалити у себе", що буде працювати завжди, "кнопка" видалити у всіх, що буде доступна при активному з'єднанні)
 (done) таск 24. перенести friends Online, friends offline в clientPartP2p
 (something went wrong...) таск 30. написать контроллер входящих сообщений (handler)
 (done) таск 27. маркировать собеседника как онлайн\оффлайн

+++ app/src/main/java/com/lackofsky/cloud_s/ui/chats/ChatDialogViewModel.kt
@@ -20,12 +20,20 @@ import com.lackofsky.cloud_s.data.database.repository.MessageRepository
 import com.lackofsky.cloud_s.data.database.repository.ReadMessageRepository
 import com.lackofsky.cloud_s.data.database.repository.UserRepository
 import com.lackofsky.cloud_s.service.ClientPartP2P
+import com.lackofsky.cloud_s.service.client.usecase.MessageRequestUseCase
+import com.lackofsky.cloud_s.service.model.MessageKey
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.firstOrNull
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -35,6 +43,7 @@ class ChatDialogViewModel @Inject constructor(private val userRepository: UserRe
                                               private val chatRepository: ChatRepository,
                                               private val chatMemberRepository: ChatMemberRepository,
                                               private val clientPartP2P: ClientPartP2P,
+    private val messageRequestUseCase: MessageRequestUseCase
                                               //private val readMessageRepository: ReadMessageRepository
     ): ViewModel() {
     private val _messages = MutableStateFlow<List<Message>?>(null)
@@ -43,16 +52,29 @@ class ChatDialogViewModel @Inject constructor(private val userRepository: UserRe
 
     //val activeFriends = MutableStateFlow<List<User?>?>(null)
     private val _activeChat = MutableStateFlow<Chat?>(null)
+    val activeChat: StateFlow<Chat?> get() = _activeChat
     val activeUserOwner = MutableStateFlow<User?>(null)
 
     private val _selectedMessages = MutableStateFlow<MutableList<Message>>(mutableListOf())
     val selectedMessages: StateFlow<MutableList<Message>> = _selectedMessages
     private val _isSelectingMode = MutableStateFlow(_selectedMessages.value.isNotEmpty())
     val isSelectingMode: StateFlow<Boolean> = _isSelectingMode
-    init{
+
+    //private val _isFriendOnline = MutableStateFlow(false)//= clientPartP2P.friendsOnline.map { friends -> friends.any { it.uniqueID == _activeChat.value!!.name } }
+    val isFriendOnline: StateFlow<Boolean> get() = clientPartP2P.friendsOnline.map {
+        friends -> friends.any{
+            it.uniqueID == _activeChat.value?.name && _activeChat.value?.type == ChatType.PRIVATE}
+    }.stateIn(viewModelScope, SharingStarted.Lazily, false)//_isFriendOnline
+
+    init {
+//        CoroutineScope(Dispatchers.IO).launch {
+//            clientPartP2P.friendsOnline.collect{ friends ->
+//                //Log.d("GrimBerry chatDialogVM", "friendsOnline: $it")
+//                _isFriendOnline.update { friends.any { it.uniqueID == _activeChat.value!!.name && _activeChat.value!!.type == ChatType.PRIVATE} }
+//            }
+//        }
 
     }
-
 //    fun setMessagesList(chatId: String){
 //        CoroutineScope(Dispatchers.IO).launch {
 //            val chat = chatRepository.getChatById(chatId)
@@ -65,6 +87,7 @@ class ChatDialogViewModel @Inject constructor(private val userRepository: UserRe
         viewModelScope.launch {
             messageRepository.getMessagesByChat(chatId).collect{
                 _messages.value = it
+                Log.d("GrimBerry chatDialogVM", "setChatId: ${it.toString()}")
             }
         }
         viewModelScope.launch {
@@ -93,41 +116,37 @@ class ChatDialogViewModel @Inject constructor(private val userRepository: UserRe
 
     }
 
-    fun sendMessage(text: String){
+    fun sendMessage(text: String){//todo change to sendPrivateMessage
         CoroutineScope(Dispatchers.IO).launch {
-            var id = 0
-            if(messages.value!!.isEmpty()){
-                id = 1
-            }else{
-                id = messages.value!!.last().id+1
-            }
-            val message = Message(
-                        id = id,
+            val chatId = _activeChat.value!!.chatId
+            val ownId = activeUserOwner.value!!.uniqueID
+            val messageToSave = Message(
                         userId = activeUserOwner.value!!.uniqueID,
-                        uniqueId = activeUserOwner.value!!.uniqueID + id,
-                chatId = activeUserOwner.value!!.uniqueID,// у власника чат айді = _activeChat.value!!.chatId, у другої сторони = айді власника
+                chatId = _activeChat.value!!.chatId ,// у власника чат айді = _activeChat.value!!.chatId, у другої сторони = айді власника
                 content = text)
-
-            val ownChatId = _activeChat.value!!.chatId
-            val messageToSave = message.copy(chatId = ownChatId)
 
 
             if(chatMembers.value.isNullOrEmpty()){
                 Log.d("GrimBerry chatDialogVM", "chatMembers.value.isNullOrEmpty()")
-                messageRepository.insertMessage(message)
+                messageRepository.insertAndUpdateMessage(messageToSave)
             }else{
                 try{
+                    val activeChat = _activeChat.value!!
                     chatMembers.value?.forEach { member ->
                         //todo - фактически данная логика поставлена на то, что все пользователи онлайн
-                        Log.d("GrimBerry chatDialogVM", message.toString())
-                        clientPartP2P.sendMessage(
-                            activeFriend = userRepository.getUserByUniqueID(member.userId).first(),
-                            message = message
-                        )
+                        // upd - наразі ця логіка для використання для one2one
+                        val client = clientPartP2P.activeFriends.value.entries
+                            .firstOrNull {
+                                it.key.uniqueID == activeChat.name && activeChat.type == ChatType.PRIVATE
+                            }?.value!!
+                        Log.d("GrimBerry chatDialogVM", "to save  "+ messageToSave.toString())
+                        val messageToSend = messageRepository.insertAndUpdateMessage( messageToSave )
+                            .copy(chatId = ownId)
+                        Log.d("GrimBerry chatDialogVM", "to send  "+ messageToSend.toString())
+                        messageRequestUseCase.sendMessageRequest(client, messageToSend)
                         Log.d("GrimBerry chatDialogVM", "send message")
                     }
 
-                    messageRepository.insertMessage( messageToSave )
                 }catch (e: Exception){
                     //додати вспливаюче повідомлення що користувач не онлайн
                     Log.d("GrimBerry chatDialogVM", e.toString())
@@ -137,12 +156,46 @@ class ChatDialogViewModel @Inject constructor(private val userRepository: UserRe
         }
     }
 
-    fun deleteMessage(message: Message):Boolean{
+    fun deleteMessage(message: Message, forOneToOne: Boolean = false):Boolean{
         //фича - удаление сообщений будет производится лишь у себя
-        CoroutineScope(Dispatchers.IO).launch{
-            messageRepository.deleteMessage(message)
+        try {
+            CoroutineScope(Dispatchers.IO).launch {
+                if(forOneToOne){
+                    //clientPartP2P.activeFriends.value
+                    val activeChat = _activeChat.value!!
+                    val client = clientPartP2P.activeFriends.value.entries
+                        .firstOrNull {
+                            it.key.uniqueID == activeChat.name && activeChat.type == ChatType.PRIVATE
+                        }?.value!!
+
+                    val isSuccess = messageRequestUseCase.deleteMessageOne2OneRequest(client,
+                        MessageKey(chatName = activeChat.name!!,
+                            contentHashCode = message.content.hashCode(),
+                            date = message.sentAt)
+                    )
+                    if(isSuccess){
+                        messageRepository.deleteMessage(message)
+                        Log.d("GrimBerry chat dialog viewModel", "deleteMessage: success")
+                    }else{
+                        Log.d("GrimBerry chat dialog viewModel", "deleteMessage: failed")
+                    }
+
+                }else{
+                    messageRepository.deleteMessage(message)
+                }
+
+            }
+        }catch (e: Exception){
+            Log.d("GrimBerry friends view model", e.toString())
+            return false
         }
         return true
+
+
+
+
+
+
     }
     fun copyToClipboard(context: Context, text: String):Boolean {
         // Получаем ClipboardManager
